jupyter repl set up w/ autoreload
