+++ grabble/__init__.py
@@ -1,1 +1,5 @@
-__version__ = '0.1.0'
+__version__ = "0.0.0"
+
+from .graphdataframe import GrabbleFrame
+
+__all__ = ["GrabbleFrame"]
