+++ notebooks/nb-env.yml
@@ -1,0 +1,6 @@
+name: grabble
+channels:
+  - conda-forge
+dependencies:
+  - python=3.8
+  - ipykernel
